fix(plc4go): stop malformed frames from livelocking or hanging the EIP/S7/KNX/BACnet drivers
- f014: validate the computed frame length against each protocol's minimum
  header size with overflow-safe arithmetic and treat violations as fatal
  framing errors, so a zero/wrapped length can no longer spin the receive
  worker forever (EIP payload+24 uint16 wrap, S7 TPKT<4, KNX header<6,
  BACnet BVLC<4 via a spoofed UDP datagram).
- f015: length-check the KNX device-descriptor response before indexing and
  make every recover/early-return path complete the result channel (and add a
  ctx.Done() case to the blocking receives) so a short response can no longer
  hang browse/read forever and leak goroutines.
- f017: recover the BACnet segmented-reassembly goroutine so a panic on a
  wire-controlled size fails the read instead of the process.

+++ plc4go/internal/eip/MessageCodec.go
@@ -90,8 +90,11 @@ func (m *MessageCodec) Receive(ctx context.Context) (spi.Message, error) {
 			// TODO: Possibly clean up ...
 			return nil, nil
 		}
-		//Second byte for the size and then add the header size 24
-		packetSize := uint32(((uint16(data[3]) << 8) + uint16(data[2])) + 24)
+		// Second byte pair is the payload size; add the 24 byte header size.
+		// The addition must happen in uint32: doing it in uint16 lets a wire
+		// length like 0xFFE8 wrap the total to 0, making Read consume nothing
+		// and the receive worker spin forever on the same unconsumed bytes.
+		packetSize := (uint32(data[3]) << 8) + uint32(data[2]) + 24
 		if num < packetSize {
 			m.log.Debug().Uint32("num", num).Uint32("packetSize", packetSize).Msg("Not enough bytes. Got: num Need: packetSize")
 			return nil, nil
